shared foldable system
